Create index.html
This is the front page of the website.  It contains the link to pie.html. It must be at least 20 lines long and contain 4 HTML elements.  We can add a css file for it (named index.html) if we'd like.  This is very primitive, but I thought we could start from here.